+++ src/nns/topology/report/tests.rs
@@ -4,9 +4,11 @@ use crate::subnet_catalog::{
     SubnetKind, SubnetSpecialization,
 };
 use crate::{
-    nns::data_center::report::NnsDataCenterRow, nns::node::report::NnsNodeRow,
-    nns::node_operator::report::NnsNodeOperatorRow, nns::node_provider::report::NnsNodeProviderRow,
-    subnet_catalog::SubnetCatalogSubnetRow,
+    nns::data_center::report::{NnsDataCenterRefreshReport, NnsDataCenterRow},
+    nns::node::report::{NnsNodeRefreshReport, NnsNodeRow},
+    nns::node_operator::report::{NnsNodeOperatorRefreshReport, NnsNodeOperatorRow},
+    nns::node_provider::report::{NnsNodeProviderRefreshReport, NnsNodeProviderRow},
+    subnet_catalog::{SubnetCatalogRefreshReport, SubnetCatalogSubnetRow},
 };
 
 #[test]